Added some changes in route guards and error page.
Added notifications to route guard, if a non-user tried to type it in
the url segment he'll be redirected to error page.
Added a link to navigate all the pathes that are not included in our
project to the error page.
Added a back to home button in the error page, and linked the logo.